added reservationDate to Ticket and changed price to long refs #22812

+++ client/src/main/java/at/ac/tuwien/inso/sepm/ticketline/client/gui/ticket/HallplanController.java
@@ -353,7 +353,7 @@ public class HallplanController implements LocalizationObserver {
                     .title(event.getTitle())
                     .build())
                 .isPaid(false)
-                .price(event.getPrice().intValue())
+                .price(event.getPrice())
                 .seat(seat)
                 .build());
         }
